requirements.txt
Django>=4.0
psycopg2-binary
djangorestframework
requests
django-ckeditor